[feature/single_live_event] refactor: move base viewmodel
