+++ feature/authentication/src/main/java/com/alxnophis/jetpack/authentication/ui/contract/AuthenticationContract.kt
@@ -4,8 +4,8 @@ import androidx.annotation.StringRes
 import arrow.optics.optics
 import com.alxnophis.jetpack.authentication.R
 import com.alxnophis.jetpack.core.base.constants.EMPTY
-import com.alxnophis.jetpack.core.base.viewmodel.UiEvent
-import com.alxnophis.jetpack.core.base.viewmodel.UiState
+import com.alxnophis.jetpack.core.ui.viewmodel.UiEvent
+import com.alxnophis.jetpack.core.ui.viewmodel.UiState
 
 internal const val NO_ERROR = 0
 
@@ -34,8 +34,8 @@ internal data class AuthenticationState(
 
     fun isFormValid(): Boolean {
         return password.isNotEmpty() &&
-            email.isNotEmpty() &&
-            (authenticationMode == AuthenticationMode.SIGN_IN || passwordRequirements.containsAll(PasswordRequirements.values().toList()))
+                email.isNotEmpty() &&
+                (authenticationMode == AuthenticationMode.SIGN_IN || passwordRequirements.containsAll(PasswordRequirements.values().toList()))
     }
 
     internal companion object {
