+++ feature/file-downloader/src/main/java/com/alxnophis/jetpack/filedownloader/ui/contract/FileDownloaderContract.kt
@@ -1,8 +1,8 @@
 package com.alxnophis.jetpack.filedownloader.ui.contract
 
 import arrow.optics.optics
-import com.alxnophis.jetpack.core.base.viewmodel.UiEvent
-import com.alxnophis.jetpack.core.base.viewmodel.UiState
+import com.alxnophis.jetpack.core.ui.viewmodel.UiEvent
+import com.alxnophis.jetpack.core.ui.viewmodel.UiState
 import com.alxnophis.jetpack.kotlin.constants.ZERO_INT
 
 internal sealed class FileDownloaderEvent : UiEvent {
