+++ feature/home/src/main/java/com/alxnophis/jetpack/home/ui/viewmodel/HomeViewModel.kt
@@ -3,7 +3,7 @@ package com.alxnophis.jetpack.home.ui.viewmodel
 import androidx.lifecycle.viewModelScope
 import arrow.core.Either
 import arrow.optics.copy
-import com.alxnophis.jetpack.core.base.viewmodel.BaseViewModel
+import com.alxnophis.jetpack.core.ui.viewmodel.BaseViewModel
 import com.alxnophis.jetpack.home.R
 import com.alxnophis.jetpack.home.domain.model.NavigationError
 import com.alxnophis.jetpack.home.domain.model.NavigationItem
