+++ feature/location-tracker/src/main/java/com/alxnophis/jetpack/location/tracker/ui/contract/LocationTrackerContract.kt
@@ -1,8 +1,8 @@
 package com.alxnophis.jetpack.location.tracker.ui.contract
 
 import arrow.optics.optics
-import com.alxnophis.jetpack.core.base.viewmodel.UiEvent
-import com.alxnophis.jetpack.core.base.viewmodel.UiState
+import com.alxnophis.jetpack.core.ui.viewmodel.UiEvent
+import com.alxnophis.jetpack.core.ui.viewmodel.UiState
 import com.alxnophis.jetpack.kotlin.constants.EMPTY
 
 internal sealed class LocationTrackerEvent : UiEvent {
