+++ feature/posts/src/main/java/com/alxnophis/jetpack/posts/ui/viewmodel/PostsViewModel.kt
@@ -3,7 +3,7 @@ package com.alxnophis.jetpack.posts.ui.viewmodel
 import androidx.lifecycle.viewModelScope
 import arrow.core.Either
 import arrow.optics.copy
-import com.alxnophis.jetpack.core.base.viewmodel.BaseViewModel
+import com.alxnophis.jetpack.core.ui.viewmodel.BaseViewModel
 import com.alxnophis.jetpack.core.ui.model.ErrorMessage
 import com.alxnophis.jetpack.posts.R
 import com.alxnophis.jetpack.posts.data.model.Post
@@ -64,16 +64,16 @@ internal class PostsViewModel(
 
     private fun PostsError.mapTo(): List<ErrorMessage> =
         currentState.errorMessages +
-            ErrorMessage(
-                id = getRandomUUID(),
-                messageId =
-                when (this@mapTo) {
-                    PostsError.Network -> R.string.posts_error_network
-                    PostsError.Server -> R.string.posts_error_server
-                    PostsError.Unknown -> R.string.posts_error_unknown
-                    PostsError.Unexpected -> R.string.posts_error_unexpected
-                }
-            )
+                ErrorMessage(
+                    id = getRandomUUID(),
+                    messageId =
+                    when (this@mapTo) {
+                        PostsError.Network -> R.string.posts_error_network
+                        PostsError.Server -> R.string.posts_error_server
+                        PostsError.Unknown -> R.string.posts_error_unknown
+                        PostsError.Unexpected -> R.string.posts_error_unexpected
+                    }
+                )
 
     private fun dismissError(errorId: Long) {
         val errorMessages = currentState.errorMessages.filterNot { it.id == errorId }
