+++ feature/settings/src/main/java/com/alxnophis/jetpack/settings/ui/viewmodel/SettingsViewModel.kt
@@ -2,7 +2,7 @@ package com.alxnophis.jetpack.settings.ui.viewmodel
 
 import androidx.lifecycle.viewModelScope
 import arrow.optics.copy
-import com.alxnophis.jetpack.core.base.viewmodel.BaseViewModel
+import com.alxnophis.jetpack.core.ui.viewmodel.BaseViewModel
 import com.alxnophis.jetpack.core.extensions.doNothing
 import com.alxnophis.jetpack.settings.ui.contract.MarketingOption
 import com.alxnophis.jetpack.settings.ui.contract.SettingsEvent
