--- shared/core/src/main/java/com/alxnophis/jetpack/core/base/viewmodel/UiState.kt
@@ -1,3 +1,0 @@
-package com.alxnophis.jetpack.core.base.viewmodel
-
-interface UiState
