+++ shared/core/src/main/java/com/alxnophis/jetpack/core/ui/viewmodel/UiEvent.kt
@@ -1,0 +1,3 @@
+package com.alxnophis.jetpack.core.ui.viewmodel
+
+interface UiEvent
